Move cf lib
